requirements.txt
# Core dependencies
click==8.1.7
sqlalchemy==2.0.23
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.3.0
python-dotenv==1.0.0

# CLI and display dependencies
tabulate==0.9.0
colorama==0.4.6

# Additional requirements for proper functionality
typing-extensions>=4.0.0
